First pass at adding SlideID buttons
Current plan is to treat Slide ID as a super filtering type that exists outside of the filtering container. I have mocked up some buttons that progress through the images, and I will be adjusting the filtering method in a moment, so I can keep all the filtering methods together.