Creating a reset function
Needed a reset function to handle when swapping between datasets. This fixes a bug that occurred if neighborhood profile settings had been made with a previous dataset, that loading a new dataset (with different columns) does not break the environment.

Also creating an early version of boolean detection and string detection. Not sure if I am going to move that elsewhere, but its a start.