Removing necessity for config files
To reduce the complexity of MAWA I am removing some old lines of code that we have made irrelevant. Previously it made sense to carry around a config file that outlined the specific data information for each project and working group, but now that we have a common frame of reference for all datasets, this sort of  design is unnecessary and confusing.

I am focusing on removing the individual code snippets first, and will remove the config files in a later commit, after I am sure that I have not broken anything.