Complete going through preprocessing step for a group and implementing all datafile processing from that, plus generalizations
Need to just go through dataset_formats.py and finish allowing a user to use the app to produce one of the dataset formats, likely Native or HALO

Upon trying to save everything using the Streamlit session management, I get a random error like  I can't seem to reliably reproduce it; it seems pretty random. When the session does save successfully, which is a majority of the time, it's very fast even though it's producing very large files (2-3 GB, so be careful saving a lot). This problem seems to be a common one but for which nobody seems to have a reliable solution