Adding prepare_df_UMAP_Plotting to SpatialUMAP
This is a function that will continue to evolve, but ultimately this is a plan to sort out only the required cells for plotting and the required feature columns. I have tested this works. Admittedly, there is a lot going on in this commit. Its confusing. More to come in the next version