Making an attempt to making areas calc faster
This is my attempt to adjust the areas calculation so that it is faster than it currently is. I am trying a method of changing the order of how the areas that are calculated from index wise to region wise. Based on preliminary results, this new method takes just a long. I am going to table this for now any maybe come back to it later.