improved deleteion of attribute from sessionstate
The previous method worked in some older versions of Python, but appears to be giving some errors in Python 2.12. I tested this in Jupyter Notebook and in the main app and so far I am not running into any errors. Theoretically this should not cause any problems, but wanted to make this a simple commit to track for the future.