Changing behavior for data_import_and_export
Previously a bunch of UI elements were removed from the Data Import and Export page that were not necessary for local use/development. I am adding them back in for a few reasons.

- It makes development of these widgets easier when developing locally
- It will force us to consider how this design will need to change when we move from the NIDAP foundry platform to the next version.

I am also populating the list of files in the ./input folder to better test a nother festure. We can undo line 281 later if its confusing